Restructure project and make the import init less of a hack
