+++ cloudpoint_server/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+clap = { version = "4.6.1", features = ["derive"] }
 axum = "0.8"
 tower-http = { version = "0.6.11", features = ["trace"] }
 tracing = "0.1.44"
