+++ cloudpoint_server/Dockerfile
@@ -7,3 +7,4 @@ FROM debian:bullseye-slim
 COPY --from=builder /clpt/target/release/cloudpoint_server /usr/local/bin/
 EXPOSE 6776
 ENTRYPOINT [ "cloudpoint_server" ]
+CMD [ "serve" ]
