+++ cloudpoint_server/compose.local.yml
@@ -29,19 +29,5 @@ services:
       db:
         condition: service_healthy
 
-  import:
-    profiles: ["import"]
-    image: dwalker109/cloudpoint
-    restart: no
-    environment:
-      V0_IMPORT: /import/data.v0
-      DATABASE_URL: postgres://postgres:postgres@db:5432/cloudpoint
-      RUST_LOG: DEBUG
-    volumes:
-      - /home/dan/mnt/data:/import/data.v0:ro
-    depends_on:
-      db:
-        condition: service_healthy
-
 volumes:
   pgdata:
