+++ cloudpoint_server/compose.prod.yml
@@ -8,6 +8,8 @@ services:
     environment:
       POSTGRES_DB: cloudpoint
       POSTGRES_PASSWORD: postgres
+    ports:
+      - "127.0.0.1:5432:5432"
     healthcheck:
       test: ["CMD-SHELL", "pg_isready -U postgres -d cloudpoint"]
       interval: 5s
@@ -30,20 +32,6 @@ services:
     environment:
       RUST_LOG: INFO
       DATABASE_URL: postgres://postgres:postgres@db:5432/cloudpoint
-    depends_on:
-      db:
-        condition: service_healthy
-
-  import:
-    profiles: ["import"]
-    image: dwalker109/cloudpoint
-    restart: no
-    environment:
-      V0_IMPORT: /import/data.v0
-      DATABASE_URL: postgres://postgres:postgres@db:5432/cloudpoint
-      RUST_LOG: DEBUG
-    volumes:
-      - /path/to/v0/data:/import/data.v0:ro
     depends_on:
       db:
         condition: service_healthy
