+++ cloudpoint_server/src/app/tests/v0.rs
@@ -16,7 +16,7 @@ const CHUNK_GZ_DATA: &[u8] = &[
 const SYNC_ITEM: &str = "000400000007E800.savedata";
 const VERSION_CID: u128 = 0x68A9918350F92B62AB8EF8AB0F2EFEA1;
 
-#[sqlx::test(fixtures("../../fixtures/chunks.sql"))]
+#[sqlx::test(fixtures("../../../fixtures/chunks.sql"))]
 async fn bad_path_segments(db_pool: Pool<sqlx::Postgres>) {
     let expected = [
         Request::head(format!("/api/v1/chunk/bad_uuid/{CHUNK_CID:032x}")).body(Body::empty()),
@@ -27,7 +27,7 @@ async fn bad_path_segments(db_pool: Pool<sqlx::Postgres>) {
         Request::put(format!("/api/v1/chunk/{USER_KEY}/bad_cid")).body(Body::empty()),
     ];
 
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     for req in expected {
         let response = app.clone().oneshot(req.unwrap()).await.unwrap();
@@ -36,9 +36,9 @@ async fn bad_path_segments(db_pool: Pool<sqlx::Postgres>) {
     }
 }
 
-#[sqlx::test(fixtures("../../fixtures/chunks.sql"))]
+#[sqlx::test(fixtures("../../../fixtures/chunks.sql"))]
 async fn chunk_head_ok(db_pool: Pool<sqlx::Postgres>) {
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     let response = app
         .oneshot(
@@ -52,9 +52,9 @@ async fn chunk_head_ok(db_pool: Pool<sqlx::Postgres>) {
     assert_eq!(response.status(), StatusCode::OK);
 }
 
-#[sqlx::test(fixtures("../../fixtures/chunks.sql"))]
+#[sqlx::test(fixtures("../../../fixtures/chunks.sql"))]
 async fn chunk_get_ok(db_pool: Pool<sqlx::Postgres>) {
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     let response = app
         .oneshot(
@@ -71,9 +71,9 @@ async fn chunk_get_ok(db_pool: Pool<sqlx::Postgres>) {
     assert_eq!(&body[..], CHUNK_GZ_DATA);
 }
 
-#[sqlx::test(fixtures("../../fixtures/chunks.sql"))]
+#[sqlx::test(fixtures("../../../fixtures/chunks.sql"))]
 async fn chunk_get_no_access_not_found(db_pool: Pool<sqlx::Postgres>) {
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     let response = app
         .oneshot(
@@ -95,7 +95,7 @@ async fn chunk_get_no_access_not_found(db_pool: Pool<sqlx::Postgres>) {
 
 #[sqlx::test] // No fixture, intentionally
 async fn chunk_put_ok(db_pool: Pool<sqlx::Postgres>) {
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     let response = app
         .clone()
@@ -126,7 +126,7 @@ async fn chunk_put_ok(db_pool: Pool<sqlx::Postgres>) {
 
 #[sqlx::test] // No fixture, intentionally
 async fn chunk_put_malformed_errors(db_pool: Pool<sqlx::Postgres>) {
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     let response = app
         .clone()
@@ -153,9 +153,9 @@ async fn chunk_put_malformed_errors(db_pool: Pool<sqlx::Postgres>) {
     assert_eq!(response.status(), StatusCode::BAD_REQUEST);
 }
 
-#[sqlx::test(fixtures("../../fixtures/versions.sql"))]
+#[sqlx::test(fixtures("../../../fixtures/versions.sql"))]
 async fn version_get_latest_ok(db_pool: Pool<sqlx::Postgres>) {
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     let response = app
         .oneshot(
@@ -189,7 +189,7 @@ async fn version_get_latest_ok(db_pool: Pool<sqlx::Postgres>) {
 
 #[sqlx::test] // No fixture, intentionally
 async fn version_get_latest_no_data(db_pool: Pool<sqlx::Postgres>) {
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     let response = app
         .oneshot(
@@ -207,9 +207,9 @@ async fn version_get_latest_no_data(db_pool: Pool<sqlx::Postgres>) {
     assert_eq!(res.get("paths").unwrap().as_array().unwrap().len(), 0);
 }
 
-#[sqlx::test(fixtures("../../fixtures/versions.sql"))]
+#[sqlx::test(fixtures("../../../fixtures/versions.sql"))]
 async fn version_get_ok(db_pool: Pool<sqlx::Postgres>) {
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     let response = app
         .oneshot(
@@ -229,9 +229,9 @@ async fn version_get_ok(db_pool: Pool<sqlx::Postgres>) {
     assert_eq!(ver.fingerprint(), VERSION_CID);
 }
 
-#[sqlx::test(fixtures("../../fixtures/versions.sql"))]
+#[sqlx::test(fixtures("../../../fixtures/versions.sql"))]
 async fn version_get_no_access_not_found(db_pool: Pool<sqlx::Postgres>) {
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     let response = app
         .oneshot(
@@ -253,7 +253,7 @@ async fn version_get_no_access_not_found(db_pool: Pool<sqlx::Postgres>) {
 
 #[sqlx::test] // No fixture, intentionally
 async fn ver_put_ok(db_pool: Pool<sqlx::Postgres>) {
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     let ver = Version::new(
         &Tree::<MemLeaf>::new(vec![], ()),
@@ -299,7 +299,7 @@ async fn ver_put_ok(db_pool: Pool<sqlx::Postgres>) {
 
 #[sqlx::test] // No fixture, intentionally
 async fn ver_put_malformed_errors(db_pool: Pool<sqlx::Postgres>) {
-    let app = super::app(AppState { db_pool });
+    let app = super::make(AppState { db_pool });
 
     let response = app
         .clone()
