+++ cloudpoint_server/src/import_v0.rs
@@ -1,17 +1,14 @@
-use anyhow::Context;
-use sqlx::PgPool;
-use std::{env, fs, path::PathBuf, str::FromStr};
-use tracing::{info, warn};
-use uuid::Uuid;
-
 use crate::{
-    HexU128,
+    hex_u128::HexU128,
     services::{chunk, version},
 };
+use anyhow::Context;
+use sqlx::PgPool;
+use std::{fs, path::Path, str::FromStr};
+use tracing::{debug, info, warn};
+use uuid::Uuid;
 
-pub async fn run(db_pool: &PgPool) -> anyhow::Result<()> {
-    let root_dir = PathBuf::from(env::var("V0_IMPORT")?);
-
+pub async fn run(db_pool: &PgPool, root_dir: &Path) -> anyhow::Result<()> {
     info!(root_dir = root_dir.to_str(), "importing v0 DUFS data");
 
     let mut chunks_count = (0_usize, 0_usize);
@@ -24,11 +21,7 @@ pub async fn run(db_pool: &PgPool) -> anyhow::Result<()> {
             let user_key = Uuid::from_str(&user_dir.file_name().unwrap().to_string_lossy())
                 .context("parsing user dir to UUID")?;
 
-            info!(
-                user_dir = user_dir.to_str(),
-                user_key = user_key.to_string(),
-                "processing user_key dir"
-            );
+            info!(user_key = user_key.to_string(), "processing user data");
 
             let chunks_path = user_dir.join("chunks");
             if !fs::exists(&chunks_path)? {
@@ -40,6 +33,8 @@ pub async fn run(db_pool: &PgPool) -> anyhow::Result<()> {
                 if entry.file_type()?.is_dir() {
                     let shard_dir = entry.path();
 
+                    debug!(shard_dir = shard_dir.to_str(), "processing shard dir");
+
                     for entry in fs::read_dir(shard_dir)? {
                         let entry = entry?;
                         if entry.file_type()?.is_file() {
@@ -47,7 +42,7 @@ pub async fn run(db_pool: &PgPool) -> anyhow::Result<()> {
                             let chunk_file = entry.path();
                             let chunk_body = fs::read(&chunk_file)?;
 
-                            info!(
+                            debug!(
                                 chunk_cid = chunk_cid.to_string(),
                                 chunk_file = chunk_file.to_str(),
                                 "processing chunk file"
@@ -59,11 +54,16 @@ pub async fn run(db_pool: &PgPool) -> anyhow::Result<()> {
                                         .await?;
 
                                     chunks_count.0 += 1;
-                                    info!("processed chunk file")
+                                    debug!("processed chunk file")
                                 }
                                 Err(err) => {
                                     chunks_count.1 += 1;
-                                    warn!(err, "skipped chunk file")
+                                    warn!(
+                                        chunk_cid = chunk_cid.to_string(),
+                                        chunk_file = chunk_file.to_str(),
+                                        err,
+                                        "skipped chunk file"
+                                    )
                                 }
                             }
                         }
@@ -82,7 +82,7 @@ pub async fn run(db_pool: &PgPool) -> anyhow::Result<()> {
                     let si_name = entry.file_name().to_owned().into_string().unwrap();
                     let si_dir = entry.path();
 
-                    info!(si_dir = si_dir.to_str(), "processing sync item dir");
+                    debug!(si_dir = si_dir.to_str(), "processing sync item dir");
 
                     for entry in fs::read_dir(si_dir)? {
                         let entry = entry?;
@@ -91,7 +91,7 @@ pub async fn run(db_pool: &PgPool) -> anyhow::Result<()> {
                             let ver_file = entry.path();
                             let ver_body = fs::read(&ver_file)?;
 
-                            info!(
+                            debug!(
                                 sync_item = si_name,
                                 ver_cid = ver_cid.to_string(),
                                 ver_file = ver_file.to_str(),
@@ -104,26 +104,32 @@ pub async fn run(db_pool: &PgPool) -> anyhow::Result<()> {
                                         .await?;
 
                                     versions_count.0 += 1;
-                                    info!("processed version file")
+                                    debug!("processed version file")
                                 }
                                 Err(err) => {
                                     versions_count.1 += 1;
-                                    warn!(err, "skipped chunk file")
+                                    warn!(
+                                        ver_cid = ver_cid.to_string(),
+                                        ver_file = ver_file.to_str(),
+                                        err,
+                                        "skipped version file"
+                                    )
                                 }
                             }
                         }
                     }
                 }
             }
-            info!(
-                chunks_ok = chunks_count.0,
-                chunks_err = chunks_count.1,
-                versions_ok = versions_count.0,
-                versions_err = versions_count.1,
-                "import finished"
-            );
         }
     }
+
+    info!(
+        chunks_ok = chunks_count.0,
+        chunks_err = chunks_count.1,
+        versions_ok = versions_count.0,
+        versions_err = versions_count.1,
+        "import finished"
+    );
 
     Ok(())
 }
